+++ src/main/java/network/crypta/runtime/endpoints/ClientContextInitParams.java
@@ -3,12 +3,12 @@ package network.crypta.runtime.endpoints;
 import java.util.Random;
 import network.crypta.client.FetchContext;
 import network.crypta.client.InsertContext;
+import network.crypta.client.async.ClientContextResources;
 import network.crypta.client.async.ClientLayerPersister;
 import network.crypta.client.async.DatastoreChecker;
 import network.crypta.client.async.USKManager;
 import network.crypta.crypt.MasterSecret;
 import network.crypta.crypt.RandomSource;
-import network.crypta.node.ClientContextResources;
 import network.crypta.node.NodeClientCore;
 import network.crypta.support.MemoryLimitedJobRunner;
 import network.crypta.support.PriorityAwareExecutor;
