+++ src/test/java/network/crypta/client/async/ClientContextTest.java
@@ -18,7 +18,6 @@ import network.crypta.config.Config;
 import network.crypta.crypt.CryptoResumeContext;
 import network.crypta.crypt.MasterSecret;
 import network.crypta.crypt.RandomSource;
-import network.crypta.node.ClientContextResources;
 import network.crypta.support.DummyJobRunner;
 import network.crypta.support.MemoryLimitedJobRunner;
 import network.crypta.support.PriorityAwareExecutor;
