+++ src/test/java/network/crypta/clients/http/QueueToadletRecommendTest.java
@@ -14,6 +14,7 @@ import network.crypta.client.InsertContext;
 import network.crypta.client.async.ClientContext;
 import network.crypta.client.async.ClientContextDefaults;
 import network.crypta.client.async.ClientContextRafFactories;
+import network.crypta.client.async.ClientContextResources;
 import network.crypta.client.async.ClientContextRuntime;
 import network.crypta.client.async.ClientContextServices;
 import network.crypta.client.async.ClientContextStorageFactories;
@@ -30,7 +31,6 @@ import network.crypta.config.Config;
 import network.crypta.crypt.MasterSecret;
 import network.crypta.crypt.RandomSource;
 import network.crypta.l10n.NodeL10n;
-import network.crypta.node.ClientContextResources;
 import network.crypta.node.DarknetPeerNode;
 import network.crypta.node.Node;
 import network.crypta.node.NodeClientCore;
