+++ src/test/java/network/crypta/runtime/core/NodeClientCoreSupportTest.java
@@ -7,6 +7,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
 import network.crypta.client.ArchiveManager;
+import network.crypta.client.async.ClientContextResources;
 import network.crypta.client.async.SimpleHealingQueue;
 import network.crypta.config.SubConfig;
 import network.crypta.crypt.DummyRandomSource;
@@ -26,7 +27,6 @@ import network.crypta.keys.Key;
 import network.crypta.keys.SSKBlock;
 import network.crypta.keys.SSKEncodeException;
 import network.crypta.keys.SSKVerifyException;
-import network.crypta.node.ClientContextResources;
 import network.crypta.node.Node;
 import network.crypta.node.NodeClientCore;
 import network.crypta.node.NodeInitException;
