+++ src/test/java/network/crypta/runtime/endpoints/fcp/LegacyQueueInsertPortTest.java
@@ -16,6 +16,7 @@ import network.crypta.client.MetadataUnresolvedException;
 import network.crypta.client.async.ClientContext;
 import network.crypta.client.async.ClientContextDefaults;
 import network.crypta.client.async.ClientContextRafFactories;
+import network.crypta.client.async.ClientContextResources;
 import network.crypta.client.async.ClientContextRuntime;
 import network.crypta.client.async.ClientContextServices;
 import network.crypta.client.async.ClientContextStorageFactories;
@@ -37,7 +38,6 @@ import network.crypta.clients.fcp.PersistentRequestClient;
 import network.crypta.config.Config;
 import network.crypta.crypt.MasterSecret;
 import network.crypta.crypt.RandomSource;
-import network.crypta.node.ClientContextResources;
 import network.crypta.node.NodeClientCore;
 import network.crypta.runtime.endpoints.ClientEndpoints;
 import network.crypta.runtime.spi.QueueBrowserUploadInsertRequest;
