+++ src/test/java/network/crypta/support/io/DelayedFreeRandomAccessBufferTest.java
@@ -9,11 +9,11 @@ import java.util.Random;
 import network.crypta.client.async.ClientContext;
 import network.crypta.client.async.ClientContextDefaults;
 import network.crypta.client.async.ClientContextRafFactories;
+import network.crypta.client.async.ClientContextResources;
 import network.crypta.client.async.ClientContextRuntime;
 import network.crypta.client.async.ClientContextServices;
 import network.crypta.client.async.ClientContextStorageFactories;
 import network.crypta.crypt.MasterSecret;
-import network.crypta.node.ClientContextResources;
 import network.crypta.support.PriorityAwareExecutor;
 import network.crypta.support.api.LockableRandomAccessBuffer.RAFLock;
 import network.crypta.support.api.LockableRandomAccessBuffer;
