+++ src/test/java/network/crypta/support/io/PersistentTempFileBucketTest.java
@@ -13,10 +13,10 @@ import java.security.SecureRandom;
 import network.crypta.client.async.ClientContext;
 import network.crypta.client.async.ClientContextDefaults;
 import network.crypta.client.async.ClientContextRafFactories;
+import network.crypta.client.async.ClientContextResources;
 import network.crypta.client.async.ClientContextRuntime;
 import network.crypta.client.async.ClientContextServices;
 import network.crypta.client.async.ClientContextStorageFactories;
-import network.crypta.node.ClientContextResources;
 import network.crypta.support.api.RandomAccessBucket;
 import network.crypta.support.api.ResumeContext;
 import org.junit.jupiter.api.AfterEach;
